Move shouldUseExpoCLI to utils
